検索画面をpageではなくモーダルコンポーネントとして使用するように実装
