--- src/app/search/layout.tsx
@@ -1,8 +1,0 @@
-import React from "react";
-import { metadata } from "./metadata";
-
-export { metadata };
-
-export default function Layout({ children }: { children: React.ReactNode }) {
-  return <>{children}</>;
-}
